pyproject.toml
[project]
name = "cct-training"
version = "0.1.0"
description = "Commitment-Constrained Training for Transformer Language Models"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "wandb",
    "pyyaml",
    "lm-eval>=0.4.0",
    "numpy",
    "tqdm",
    "safetensors",
    "scikit-learn",
]

[project.optional-dependencies]
multi-gpu = ["deepspeed>=0.12.0"]
dev = ["pytest", "black", "ruff"]
